Windows.

+++ src/world/Chunk.hpp
@@ -6,7 +6,7 @@
 #include <sys\types.h>
 #endif
 
-const static uint16_t CHUNK_SIZE = 256;
+const static int CHUNK_SIZE = 256;
 
 class Chunk : public sf::Drawable, public sf::Transformable
 {
